vid-engine/requirements.txt
yt-dlp>=2024.1.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
